moved cuts
