+++ examples/mcsp/dp_mcsp.py
@@ -14,7 +14,7 @@ from pyomo.environ import (
     value,
 )
 from pyodsp.solver.pyomo_solver import PyomoSolver, SolverConfig
-from pyodsp.alg.cuts import OptimalityCut
+from pyodsp.alg.bm.cuts import OptimalityCut
 from pyodsp.alg.bm.cuts_manager import CutInfo
 from pyodsp.dec.dd.message import DdFinalDnMessage
 from pyodsp.dec.dd.mip_heuristic_root import IMipHeuristicRoot
