+++ pyodsp/alg/bm/bm.py
@@ -10,7 +10,7 @@ from pyomo.environ import Var, ScalarVar, Reals, RangeSet
 from pyodsp.solver.pyomo_solver import PyomoSolver
 from pyodsp.solver.pyomo_utils import add_terms_to_objective
 
-from ..cuts import CutList
+from .cuts import CutList
 from .cuts_manager import CutInfo
 from .logger import BmLogger
 from .cp import CuttingPlaneMethod
