+++ pyodsp/alg/bm/cp.py
@@ -4,7 +4,7 @@ from pyomo.environ import ScalarVar, Constraint
 
 from pyodsp.solver.pyomo_solver import PyomoSolver
 from .cuts_manager import CutsManager, CutInfo
-from ..cuts import CutList, OptimalityCut, FeasibilityCut
+from .cuts import CutList, OptimalityCut, FeasibilityCut
 
 from ..params import BM_ABS_TOLERANCE
