+++ pyodsp/dec/dd/mip_heuristic_root.py
@@ -14,7 +14,7 @@ from pyomo.environ import (
     value,
 )
 from pyodsp.solver.pyomo_solver import PyomoSolver, SolverConfig
-from pyodsp.alg.cuts import OptimalityCut
+from pyodsp.alg.bm.cuts import OptimalityCut
 from pyodsp.alg.bm.cuts_manager import CutInfo
 from .message import DdFinalDnMessage
